+++ src/TrashLib.Tests/CustomFormat/Processors/GuideSteps/ConfigStepTest.cs
@@ -4,7 +4,6 @@ using NUnit.Framework;
 using TestLibrary.AutoFixture;
 using TrashLib.Config.Services;
 using TrashLib.Services.CustomFormat.Models;
-using TrashLib.Services.CustomFormat.Models.Cache;
 using TrashLib.Services.CustomFormat.Processors.GuideSteps;
 using TrashLib.TestLibrary;
 
@@ -15,51 +14,19 @@ namespace TrashLib.Tests.CustomFormat.Processors.GuideSteps;
 public class ConfigStepTest
 {
     [Test, AutoMockData]
-    public void Cache_names_are_used_instead_of_name_in_json_data(ConfigStep processor)
-    {
-        var testProcessedCfs = new List<ProcessedCustomFormatData>
-        {
-            NewCf.Processed("name1", "id1", 100),
-            NewCf.Processed("name3", "id3", new TrashIdMapping("id3", "name1"))
-        };
-
-        var testConfig = new CustomFormatConfig[]
-        {
-            new()
-            {
-                Names = new List<string> {"name1"}
-            }
-        };
-
-        processor.Process(testProcessedCfs, testConfig);
-
-        processor.CustomFormatsNotInGuide.Should().BeEmpty();
-        processor.ConfigData.Should().BeEquivalentTo(new List<ProcessedConfigData>
-        {
-            new()
-            {
-                CustomFormats = new List<ProcessedCustomFormatData>
-                    {testProcessedCfs[1]}
-            }
-        }, op => op
-            .Using<JToken>(jctx => jctx.Subject.Should().BeEquivalentTo(jctx.Expectation))
-            .WhenTypeIs<JToken>());
-    }
-
-    [Test, AutoMockData]
     public void Custom_formats_missing_from_config_are_skipped(ConfigStep processor)
     {
         var testProcessedCfs = new List<ProcessedCustomFormatData>
         {
-            NewCf.Processed("name1", ""),
-            NewCf.Processed("name2", "")
+            NewCf.Processed("name1", "id1"),
+            NewCf.Processed("name2", "id2")
         };
 
         var testConfig = new CustomFormatConfig[]
         {
             new()
             {
-                Names = new List<string> {"name1"}
+                TrashIds = new List<string> {"id1"}
             }
         };
 
@@ -72,7 +39,7 @@ public class ConfigStepTest
             {
                 CustomFormats = new List<ProcessedCustomFormatData>
                 {
-                    NewCf.Processed("name1", "")
+                    NewCf.Processed("name1", "id1")
                 }
             }
         }, op => op
@@ -85,21 +52,21 @@ public class ConfigStepTest
     {
         var testProcessedCfs = new List<ProcessedCustomFormatData>
         {
-            NewCf.Processed("name1", ""),
-            NewCf.Processed("name2", "")
+            NewCf.Processed("name1", "id1"),
+            NewCf.Processed("name2", "id2")
         };
 
         var testConfig = new CustomFormatConfig[]
         {
             new()
             {
-                Names = new List<string> {"name1", "name3"}
+                TrashIds = new List<string> {"id1", "id3"}
             }
         };
 
         processor.Process(testProcessedCfs, testConfig);
 
-        processor.CustomFormatsNotInGuide.Should().BeEquivalentTo(new List<string> {"name3"}, op => op
+        processor.CustomFormatsNotInGuide.Should().BeEquivalentTo(new List<string> {"id3"}, op => op
             .Using<JToken>(jctx => jctx.Subject.Should().BeEquivalentTo(jctx.Expectation))
             .WhenTypeIs<JToken>());
         processor.ConfigData.Should().BeEquivalentTo(new List<ProcessedConfigData>
@@ -108,7 +75,7 @@ public class ConfigStepTest
             {
                 CustomFormats = new List<ProcessedCustomFormatData>
                 {
-                    NewCf.Processed("name1", "")
+                    NewCf.Processed("name1", "id1")
                 }
             }
         }, op => op
@@ -117,7 +84,7 @@ public class ConfigStepTest
     }
 
     [Test, AutoMockData]
-    public void Duplicate_config_name_and_id_are_ignored(ConfigStep processor)
+    public void Duplicate_config_trash_ids_are_ignored(ConfigStep processor)
     {
         var testProcessedCfs = new List<ProcessedCustomFormatData>
         {
@@ -126,11 +93,7 @@ public class ConfigStepTest
 
         var testConfig = new CustomFormatConfig[]
         {
-            new()
-            {
-                Names = new List<string> {"name1"},
-                TrashIds = new List<string> {"id1"}
-            }
+            new() {TrashIds = new List<string> {"id1", "id1"}}
         };
 
         processor.Process(testProcessedCfs, testConfig);
@@ -143,68 +106,5 @@ public class ConfigStepTest
                 CustomFormats = new List<ProcessedCustomFormatData> {testProcessedCfs[0]}
             }
         });
-    }
-
-    [Test, AutoMockData]
-    public void Duplicate_config_names_are_ignored(ConfigStep processor)
-    {
-        var testProcessedCfs = new List<ProcessedCustomFormatData>
-        {
-            NewCf.Processed("name1", "id1")
-        };
-
-        var testConfig = new CustomFormatConfig[]
-        {
-            new() {Names = new List<string> {"name1", "name1"}}
-        };
-
-        processor.Process(testProcessedCfs, testConfig);
-
-        processor.CustomFormatsNotInGuide.Should().BeEmpty();
-        processor.ConfigData.Should().BeEquivalentTo(new List<ProcessedConfigData>
-        {
-            new()
-            {
-                CustomFormats = new List<ProcessedCustomFormatData> {testProcessedCfs[0]}
-            }
-        });
-    }
-
-    [Test, AutoMockData]
-    public void Find_custom_formats_by_name_and_trash_id(ConfigStep processor)
-    {
-        var testProcessedCfs = new List<ProcessedCustomFormatData>
-        {
-            NewCf.Processed("name1", "id1", 100),
-            NewCf.Processed("name3", "id3"),
-            NewCf.Processed("name4", "id4")
-        };
-
-        var testConfig = new CustomFormatConfig[]
-        {
-            new()
-            {
-                Names = new List<string> {"name1", "name3"},
-                TrashIds = new List<string> {"id1", "id4"},
-                QualityProfiles = new List<QualityProfileConfig>
-                {
-                    new() {Name = "profile1", Score = 50}
-                }
-            }
-        };
-
-        processor.Process(testProcessedCfs, testConfig);
-
-        processor.CustomFormatsNotInGuide.Should().BeEmpty();
-        processor.ConfigData.Should().BeEquivalentTo(new List<ProcessedConfigData>
-        {
-            new()
-            {
-                CustomFormats = testProcessedCfs,
-                QualityProfiles = testConfig[0].QualityProfiles
-            }
-        }, op => op
-            .Using<JToken>(jctx => jctx.Subject.Should().BeEquivalentTo(jctx.Expectation))
-            .WhenTypeIs<JToken>());
     }
 }