+++ src/TrashLib.Tests/GlobalTestSetup.cs
@@ -1,0 +1,14 @@
+using System.Diagnostics.CodeAnalysis;
+using FluentAssertions;
+using NUnit.Framework;
+
+[SetUpFixture]
+[SuppressMessage("ReSharper", "CheckNamespace")]
+public class GlobalTestSetup
+{
+    [OneTimeSetUp]
+    public void Setup()
+    {
+        AssertionOptions.FormattingOptions.MaxDepth = 100;
+    }
+}