+++ src/TrashLib/Services/CustomFormat/Models/Cache/CustomFormatCache.cs
@@ -14,14 +14,12 @@ public class CustomFormatCache
 
 public class TrashIdMapping
 {
-    public TrashIdMapping(string trashId, string customFormatName, int customFormatId = default)
+    public TrashIdMapping(string trashId, int customFormatId = default)
     {
-        CustomFormatName = customFormatName;
         TrashId = trashId;
         CustomFormatId = customFormatId;
     }
 
-    public string CustomFormatName { get; set; }
     public string TrashId { get; }
     public int CustomFormatId { get; set; }
 }