+++ src/TrashLib/Services/CustomFormat/Processors/GuideSteps/ConfigStep.cs
@@ -1,51 +1,30 @@
 using Common.Extensions;
-using Serilog;
 using TrashLib.Config.Services;
 using TrashLib.Services.CustomFormat.Models;
 
 namespace TrashLib.Services.CustomFormat.Processors.GuideSteps;
 
+/// <remarks>
+/// The purpose of this step is to validate the custom format data in the configs:
+///
+/// - Validate that custom formats specified in the config exist in the guide.
+/// - Removal of duplicates.
+/// </remarks>
 public class ConfigStep : IConfigStep
 {
-    private readonly ILogger _log;
     private readonly List<ProcessedConfigData> _configData = new();
     private readonly List<string> _customFormatsNotInGuide = new();
 
     public IReadOnlyCollection<string> CustomFormatsNotInGuide => _customFormatsNotInGuide;
     public IReadOnlyCollection<ProcessedConfigData> ConfigData => _configData;
 
-    public ConfigStep(ILogger log)
-    {
-        _log = log;
-    }
-
     public void Process(
         IReadOnlyCollection<ProcessedCustomFormatData> processedCfs,
         IReadOnlyCollection<CustomFormatConfig> config)
     {
-        if (config.SelectMany(x => x.Names).Any())
-        {
-            _log.Warning(
-                "`names` list for `custom_formats` is deprecated and will be removed in the future; use " +
-                "`trash_ids` instead");
-        }
-
         foreach (var singleConfig in config)
         {
             var validCfs = new List<ProcessedCustomFormatData>();
-
-            foreach (var name in singleConfig.Names)
-            {
-                var match = FindCustomFormatByName(processedCfs, name);
-                if (match == null)
-                {
-                    _customFormatsNotInGuide.Add(name);
-                }
-                else
-                {
-                    validCfs.Add(match);
-                }
-            }
 
             foreach (var trashId in singleConfig.TrashIds)
             {
@@ -68,12 +47,5 @@ public class ConfigStep : IConfigStep
                     .ToList()
             });
         }
-    }
-
-    private static ProcessedCustomFormatData? FindCustomFormatByName(
-        IReadOnlyCollection<ProcessedCustomFormatData> processedCfs, string name)
-    {
-        return processedCfs.FirstOrDefault(cf => cf.CacheEntry?.CustomFormatName.EqualsIgnoreCase(name) ?? false)
-               ?? processedCfs.FirstOrDefault(cf => cf.Name.EqualsIgnoreCase(name));
     }
 }