+++ src/TrashLib/Services/CustomFormat/Processors/GuideSteps/ICustomFormatStep.cs
@@ -8,8 +8,6 @@ public interface ICustomFormatStep
 {
     IReadOnlyCollection<ProcessedCustomFormatData> ProcessedCustomFormats { get; }
     IReadOnlyCollection<TrashIdMapping> DeletedCustomFormatsInCache { get; }
-    IReadOnlyCollection<(string, string)> CustomFormatsWithOutdatedNames { get; }
-    IDictionary<string, List<ProcessedCustomFormatData>> DuplicatedCustomFormats { get; }
 
     void Process(IList<CustomFormatData> customFormatGuideData,
         IReadOnlyCollection<CustomFormatConfig> config, CustomFormatCache? cache);